EOS-17542: Rollback API refinement
